copy over missing changes to cargo rocket
in #4 launch controller now teleports the rocket to the destination instead of having it to actually fly around. some changes was made to the rocket class as well. however the original author did not notice that cargo rocket is on another branch of the class hierarchy and did not adjust cargo rocket class. this commit copies over those changes to cargo rocket class